tweaks

+++ examples/validation/validate_image2.py
@@ -35,7 +35,7 @@ data = np.array(
     np.uint8,
 )
 
-for dy, dtype in enumerate(["float32", "uint16"]):
+for dy, dtype in enumerate(["float32", "uint32"]):
     typed_data = data.astype(dtype)
 
     for dx, interpolation in enumerate(["nearest", "linear", "cubic"]):
